fix: 2가지 일을 하는 reducer(유저 인풋값 임시저장하는 리듀서), 단일책임하도록 수정 & type위치 이동
